Refactored compilation
