+++ external-crates/move/crates/move-analyzer/src/code_action.rs
@@ -8,8 +8,8 @@ use crate::{
     },
     context::Context,
     symbols::{
+        compilation::get_compiled_pkg,
         cursor::{ChainInfo, CursorContext},
-        get_compiled_pkg,
         runner::SymbolicatorRunner,
         types::{CompiledPkgInfo, PrecomputedPkgInfo, Symbols},
     },
