+++ external-crates/move/crates/move-analyzer/tests/ide_testsuite.rs
@@ -16,8 +16,8 @@ use move_analyzer::{
     completions::{compute_completions_with_symbols, utils::compute_cursor},
     inlay_hints::inlay_hints_internal,
     symbols::{
+        compilation::get_compiled_pkg,
         compute_symbols, compute_symbols_parsed_program, compute_symbols_pre_process,
-        get_compiled_pkg,
         requests::{def_info_doc_string, maybe_convert_for_guard},
         types::{CompiledPkgInfo, Symbols, SymbolsComputationData, UseDefMap},
     },
